+++ R/Spectra.R
@@ -383,6 +383,13 @@ NULL
 #'       `suffix.y`. This is to avoid modifying any core spectra
 #'       variables that would lead to an invalid object.
 #'
+#'    - Duplicated Spectra keys (i.e. `x[[by.x]]`) are not
+#'      allowed. Duplicated keys in the `DataFrame` (i.e `y[[by.y]]`)
+#'      throw a warning and only the last occurrence is kept. These
+#'      should be explored and ideally be removed using for
+#'      `QFeatures::reduceDataFrame()`, `PMS::reducePSMs()` or similar
+#'      functions.
+#'
 #' Several `Spectra` objects can be concatenated into a single object with the
 #' `c` or the `concatenateSpectra` function. Concatenation will fail if the
 #' processing queue of any of the `Spectra` objects is not empty or if
@@ -935,12 +942,13 @@ NULL
 #'
 #'
 #' ## Adding new spectra variables
-#' spv <- DataFrame(spectrumId = sciex$spectrumId[3:12], ## used for merging
+#' sciex1 <- filterDataOrigin(sciex, dataOrigin(sciex)[1])
+#' spv <- DataFrame(spectrumId = sciex1$spectrumId[3:12], ## used for merging
 #'                  var1 = rnorm(10),
 #'                  var2 = sample(letters, 10))
 #' spv
 #'
-#' sciex2 <- joinSpectraData(sciex, spv, by.y = "spectrumId")
+#' sciex2 <- joinSpectraData(sciex1, spv, by.y = "spectrumId")
 #'
 #' spectraVariables(sciex2)
 #' spectraData(sciex2)[1:13, c("spectrumId", "var1", "var2")]